Where to put state update logic
As long as synchronous side effect free code is concerned it would be technically fine to carry on with handleAddToCart state update logic and outsource it to be reusable in CartItem component so that addItem action creator becomes useless.
The problem is it would defeat the purpose of Redux to centralize state management because when faced with the dilemma of Fat Reducer vs Fat Action Creator vs Fat Component, you should reserve Reducers to synchronous pure function.