Manage Http states with Redux
There are two points to notice:
It would have been perfectly fine to dispatch three different action types, opt for Fat reducers by keeping there state update logic in Redux. In this case Fat component is prefered to display in the component the asynchronous logic.
Secondly, it would have been perfectly fine not to drill props down to the Notification component but instead to lift Redux connection down this component.